[REF] Add mozaik_payment: extract common logic for paying memberships and donations
