+++ mozaik_membership_payment/__manifest__.py
@@ -10,10 +10,13 @@
     "author": "ACSONE SA/NV",
     "website": "https://github.com/mozaik-association/mozaik",
     "depends": [
+        # Mozaik
         "mozaik_membership",
         "mozaik_membership_request",
         "mozaik_membership_request_autovalidate",
         "mozaik_account",
+        "mozaik_payment",
+        # Odoo
         "account",
         "payment",
     ],
