+++ mozaik_payment/__manifest__.py
@@ -1,0 +1,18 @@
+# Copyright 2023 ACSONE SA/NV
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+{
+    "name": "Mozaik Payment",
+    "summary": """
+        Adds Mozaik-specific features for paying""",
+    "version": "14.0.1.0.0",
+    "license": "AGPL-3",
+    "author": "ACSONE SA/NV",
+    "website": "https://github.com/mozaik-association/mozaik",
+    "depends": [
+        "payment",
+    ],
+    "data": [
+        "views/assets.xml",
+    ],
+}
