+++ mozaik_payment/views/assets.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <template id="assets_frontend" inherit_id="web.assets_frontend">
+        <xpath expr="script[last()]" position="after">
+            <script
+                type="text/javascript"
+                src="/mozaik_payment/static/src/js/payment_form.js"
+            />
+        </xpath>
+    </template>
+</odoo>
