+++ setup/mozaik_payment/odoo/addons/mozaik_payment
@@ -1,0 +1,1 @@
+../../../../mozaik_payment
